refactor layout of finances page
